+++ extensions/debug/views/default/toolbar.php
@@ -5,7 +5,7 @@
  * @var string $tag
  * @var string $position
  */
-use yii\helpers\Html;
+use yii\helpers\Url;
 
 $minJs = <<<EOD
 document.getElementById('yii-debug-toolbar').style.display = 'none';
@@ -28,7 +28,7 @@ $url = $firstPanel->getUrl();
 ?>
 <div id="yii-debug-toolbar" class="yii-debug-toolbar-<?= $position ?>">
     <div class="yii-debug-toolbar-block title">
-        <a href="<?= Html::url(['index']) ?>">
+        <a href="<?= Url::to(['index']) ?>">
             <img width="29" height="30" alt="" src="<?= \yii\debug\Module::getYiiLogo() ?>">
             Yii Debugger
         </a>