+++ extensions/smarty/ViewRenderer.php
@@ -12,8 +12,8 @@ namespace yii\smarty;
 use Yii;
 use Smarty;
 use yii\base\View;
-use yii\helpers\Html;
 use yii\base\ViewRenderer as BaseViewRenderer;
+use yii\helpers\Url;
 
 /**
  * SmartyViewRenderer allows you to use Smarty templates in views.
@@ -70,7 +70,7 @@ class ViewRenderer extends BaseViewRenderer
 		array_unshift($params, $params['route']) ;
 		unset($params['route']);
 
-		return Html::url($params);
+		return Url::to($params);
 	}
 
 	/**